# testdata/server/openapi.yaml
openapi: 3.1.0
info:
  title: Kite agent server
  version: "1"
  description: >
    HTTP+JSON+SSE frontend for kite serve. One app.Controller per session;
    turns stream as Server-Sent Events. See
    docs/specs/2026-09-03-agent-server-design.md.
servers:
  - url: http://kite
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: >
        Required on every /v1/ path when kite serve listens on TCP
        (--listen): the per-process token printed at startup, sent as
        "Authorization: Bearer <token>". Missing or wrong tokens get 401 with
        WWW-Authenticate: Bearer. The Unix-socket listener has no token.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
    Usage:
      type: object
      properties:
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        cached_input_tokens:
          type: integer
...
    Sandbox:
      type: object
      properties:
        mode:
          type: string
        network:
          type: string
        bash_available:
          type: boolean
        summary:
          type: string
...
    Compaction:
      type: object
      description: The compaction_completed event payload, also returned by POST /v1/sessions/{id}/compact.
      properties:
        checkpoint_id:
          type: string
        reason:
          type: string
          enum: [manual, threshold, overflow]
        tokens_before:
          type: integer
        estimated_tokens_after:
          type: integer
        automatic:
          type: boolean
        usage:
          $ref: "#/components/schemas/Usage"
        noop:
          type: boolean
      required: [reason, tokens_before, estimated_tokens_after, automatic, noop]
...
  parameters:
    sessionID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
paths:
...
  /v1/sessions/{id}/compact:
    parameters:
      - $ref: "#/components/parameters/sessionID"
    post:
      summary: Compact the session's context now
      description: >
        Runs one manual compaction and returns its result once it finishes;
        no events are streamed. Refused while a turn is running, and turns
        are refused while the compaction runs. Closing the request cancels
        the compaction.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                focus:
                  type: string
                  description: Optional guidance for what the summary should preserve
      responses:
        "200":
          description: >
            Compaction finished (noop:true when there was nothing to compact).
            The body is empty when the compaction was cancelled by process
            shutdown, since there is no result to report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compaction"
        "400":
          description: Invalid JSON body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Session not open
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: turn_active (a turn or compaction is running) or compaction_failed (the previous context is still in effect)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/sandbox/reload:
    post:
      summary: Re-read the sandbox configuration and apply it to the running process
      description: >
        Applies the current [sandbox] configuration to the running process
        without restarting kite. read_paths, network, and driver take effect;
        allow_env changes and a sandbox that was unavailable at startup still
        require a restart and are reported as a failure. One sandbox serves
        every session, so the reload is refused while any session has a turn in
        flight.
      responses:
        "200":
          description: The sandbox state now in effect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "409":
          description: A turn is active, or the reload failed and the previous sandbox is still in effect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "501":
          description: Sandbox reload is not available in this process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...